# resolveTokenValues perf improvements