--- a/benchmark/stats/applyToDocument.js.json
+++ b/benchmark/stats/applyToDocument.js.json
@@ -1,1 +1,1 @@
-{"rawDuration":2058.162999998778,"duration":1.1080457964165424,"input":"applyToDocument.js","output":"output\\applyToDocument.js\\flamegraph.html"}
+{"rawDuration":3291.5448000133038,"duration":1.2571321609905253,"input":"applyToDocument.js","output":"output\\applyToDocument.js\\flamegraph.html"}