--- a/src/plugin/generateTokensToCreate.ts
+++ b/src/plugin/generateTokensToCreate.ts
@@ -2,14 +2,15 @@ import { TokenSetStatus } from '@/constants/TokenSetStatus';
 import { TokenTypes } from '@/constants/TokenTypes';
 import { ThemeObject } from '@/types';
 import { AnyTokenList } from '@/types/tokens';
-import { mergeTokenGroups, resolveTokenValues } from '@/utils/tokenHelpers';
+import { defaultTokenResolver } from '@/utils/TokenResolver';
+import { mergeTokenGroups } from '@/utils/tokenHelpers';
 
 export function generateTokensToCreate(theme: ThemeObject, tokens: Record<string, AnyTokenList>, availableTokenTypes: TokenTypes[]) {
   // Big O(resolveTokenValues * mergeTokenGroups)
   const enabledTokenSets = Object.entries(theme.selectedTokenSets)
     .filter(([, status]) => status === TokenSetStatus.ENABLED)
     .map(([tokenSet]) => tokenSet);
-  const resolved = resolveTokenValues(mergeTokenGroups(tokens, theme.selectedTokenSets));
+  const resolved = defaultTokenResolver.setTokens(mergeTokenGroups(tokens, theme.selectedTokenSets));
   return resolved.filter(
     (token) => ((!token.internal__Parent || enabledTokenSets.includes(token.internal__Parent)) && availableTokenTypes.includes(token.type)), // filter out SOURCE tokens
   );