--- a/src/types/TokenMap.ts
+++ b/src/types/TokenMap.ts
@@ -1,0 +1,3 @@
+import { SingleToken } from '@/types/tokens';
+
+export type TokenMap = Map<string, SingleToken['value']>;