--- a/src/utils/getPathName.ts
+++ b/src/utils/getPathName.ts
@@ -1,0 +1,6 @@
+export function getPathName(ref: string) {
+  if (ref.startsWith('{')) {
+    return ref.slice(1, ref.length - 1);
+  }
+  return ref.substring(1);
+}