Fixes https://github.com/tokens-studio/figma-plugin/issues/1755
Fixes https://github.com/tokens-studio/figma-plugin/issues/2132

This pull request involves major changes to the token resolution system, including the addition of a new `TokenResolver` class and the replacement of the old `resolveTokenValues` function with a new `defaultTokenResolver` utility. These changes were made to improve performance and simplify the code. Additionally, some files were modified to remove unused code and simplify logic. It's able to handle deeply nested token sets performantly, as we no longer iterate over all remaining references.

* `src/utils/TokenResolver.ts`: Added a new `TokenResolver` class to resolve tokens and their references efficiently and flexibly in `src/utils/TokenResolver.ts`. (<a href="diffhunk://#diff-b4809e6851ac3bf8ff55467e9606b901d782180f05572177646b5b87641099e8R1-R267">src/utils/TokenResolver.tsR1-R267</a>)
* `src/utils/TokenResolver.test.ts`: Renamed `src/utils/tokenHelpers.test.ts` to `src/utils/TokenResolver.test.ts` and added tests to ensure the `TokenResolver` class can handle complex token structures without causing stack overflows. Also added new tokens and updated imports. (<a href="diffhunk://#diff-0a3339bdfaf2a5f4e6713b9b421551a456db51d5ffe9c106a236eb565cf791e7L440-R502">src/utils/TokenResolver.test.tsL440-R502</a>)
* `src/utils/tokenHelpers.ts`: Simplified and optimized `tokenHelpers.ts` by removing unused functions and imports and replacing them with the new `defaultTokenResolver` utility for more efficient token resolution. (F4d6c2e6L2R2)
* `src/app/store/useTokens.tsx`: Replaced the `resolveTokenValues` function with the new `defaultTokenResolver` in `useTokens.tsx` for improved performance and simplified code. (PR #X)
* `src/utils/alias/getAliasValue.ts`: Removed logic and comment related to `typography` and `boxShadow` values in `getAliasValue` function to simplify the code. (src/utils/alias/getAliasValue.ts) 


## Benchmarks
Done on M2 Max Pro
| Description                                            | Before   | After    |
|--------------------------------------------------------|----------|----------|
| Resolving test tokens (from TokenResolver.test.ts)     | 22ms     | 23ms     |
| Resolving tokens that have a depth of 150              | 3269ms   | 10.6ms   |